+++ packages/data-types/src/index.ts
@@ -83,3 +83,5 @@ export * from './lib/types/item-entry.type';
 export * from './lib/types/shop-names.type';
 export * from './lib/types/requirement-entry.type';
 export * from './lib/types/calendar-event.type';
+export * from './lib/types/minimal-npc.type';
+export * from './lib/types/gifting-preference-keys.type';