+++ packages/guide/src/app/npcs/components/gifting-grid/gifting-grid.component.ts
@@ -1,6 +1,6 @@
 import { Component, HostBinding, Input, ViewEncapsulation } from '@angular/core';
-import { UiIcon } from "../../../shared/enums/ui-icon.enum";
 import { GiftPreferences } from "@ci/data-types";
+import { preferencesMap } from "../../../shared/constants/preference-map.const";
 
 @Component({
     selector: 'app-gifting-grid',
@@ -12,16 +12,5 @@ export class GiftingGridComponent {
     @HostBinding('class.gifting-preference-grid') setCssClass = true;
     @Input({required: true}) preferences!: GiftPreferences
 
-    protected preferencesMap: {
-        icon: UiIcon,
-        label: string;
-        preferenceField: string
-    }[] = [
-        {icon: UiIcon.LOVE, label: 'Favorite', preferenceField: 'favoritePreferences'},
-        {icon: UiIcon.LOVE, label: 'Love', preferenceField: 'lovePreferences'},
-        {icon: UiIcon.LIKE, label: 'Like', preferenceField: 'likePreferences'},
-        {icon: UiIcon.NEUTRAL, label: 'Neutral', preferenceField: 'neutralPreferences'},
-        {icon: UiIcon.DISLIKE, label: 'Dislike', preferenceField: 'dislikePreferences'},
-        {icon: UiIcon.HATE, label: 'Hate', preferenceField: 'hatePreferences'},
-    ];
+    protected preferencesMap = preferencesMap
 }